Change the Animation system to use Long instead of Floats in calculating time for extra precision

+++ app/src/main/java/com/arthurgichuhi/kotlinopengl/core/animation/animation/KeyFrame.kt
@@ -3,6 +3,6 @@ package com.arthurgichuhi.kotlinopengl.core.animation.animation
 import de.javagl.jgltf.model.NodeModel
 
 data class KeyFrame(
-    var time:Float = 0f,
+    var time:Long = 0,
     var boneTransforms: MutableMap<NodeModel,BoneTransform> = HashMap()
 )
